Proposed Changes
For lines 399, 413, 427, 442, and 456 I entered in a centering code, not sure if it works.  Also, I added an "i" to Elikem's last name on line 402.  I centered the "Contact" nameplate on the page on line 491. I also may have messed up on line 499, I accidentally deleted something and tried to put it back, not sure if it is the same code. Elikem also requests that the nameplate boxes fill the page which I did not do yet.